fix(*) rename encryption -> bcrypt
